Wc: Move tests for it's own file
